third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/embed-tag-control-char-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?q=%3Cembed%20name=%27plugin%27%20type=%27application/x-blink-test-plugin%27%20src=%27http://127.0.0.1:8000/security/xssAuditor/resources/dummy.swf%05%27%3E%3C/embed%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
